+++ clang/lib/Analysis/LifetimeSafety/CMakeLists.txt
@@ -1,0 +1,12 @@
+add_clang_library(clangAnalysisLifetimeSafety
+  Checker.cpp
+  Facts.cpp
+  FactsGenerator.cpp
+  LifetimeAnnotations.cpp
+  LifetimeSafety.cpp
+  LiveOrigins.cpp
+  Loans.cpp
+  LoanPropagation.cpp
+  Origins.cpp
+  )
+
